# market-web/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Market Web REST API
  description: REST API для витрины магазина мячей
  version: 1.0.0
  contact:
    name: Margarita
    email: rl-c@yandex.ru

servers:
  - url: http://localhost:8080
    description: Development server

tags:
  - name: products
    description: Операции с товарами
  - name: orders
    description: Операции с заказами

paths:
  /items:
    get:
      tags:
        - products
      summary: Получить список товаров
      operationId: getProducts
      parameters:
        - name: search
          in: query
          description: Поиск по названию или описанию
          required: false
          schema:
            type: string
        - name: sort
          in: query
          description: Сортировка
          required: false
          schema:
            type: string
            enum: [NO, PRICE_ASC, PRICE_DESC, NAME_ASC, NAME_DESC]
            default: NO
        - name: pageNumber
          in: query
          description: Номер страницы (начиная с 1)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            enum: [2, 5, 10, 20, 50, 100]
            default: 5
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'

    post:
      tags:
        - products
      summary: Создать новый товар
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Ошибка валидации

  /items/{id}:
    get:
      tags:
        - products
      summary: Получить товар по ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          description: ID товара
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Товар найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Товар не найден

  /cart/items:
    get:
      tags:
        - products
      summary: Получить корзину
      operationId: getCart
      responses:
        '200':
          description: Содержимое корзины
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartItem'

    post:
      tags:
        - products
      summary: Изменить количество товара в корзине
      operationId: updateCartItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartRequest'
      responses:
        '200':
          description: Корзина обновлена
        '404':
          description: Товар не найден

  /orders:
    get:
      tags:
        - orders
      summary: Получить список заказов
      operationId: getOrders
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

    post:
      tags:
        - orders
      summary: Создать заказ из корзины
      operationId: createOrder
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateOrderResponse'
        '400':
          description: Корзина пуста

  /orders/{id}:
    get:
      tags:
        - orders
      summary: Получить заказ по ID
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          description: ID заказа
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Заказ найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Заказ не найден

components:
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID товара
          example: 1
        title:
          type: string
          description: Название товара
          example: Мяч футбольный
        description:
          type: string
          description: Описание товара
          example: Мяч для игры в футбол
        image:
          type: string
          description: URL изображения
          example: /images/football.jpg
        price:
          type: integer
          format: int64
          description: Цена
          example: 1000
        cartCount:
          type: integer
          description: Количество в корзине
          example: 0

    ProductPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        totalElements:
          type: integer
          format: int64
          description: Всего элементов
          example: 42
        totalPages:
          type: integer
          description: Всего страниц
          example: 9
        size:
          type: integer
          description: Размер страницы
          example: 5
        number:
          type: integer
          description: Номер текущей страницы (начиная с 0)
          example: 0
        hasPrevious:
          type: boolean
          description: Есть ли предыдущая страница
          example: false
        hasNext:
          type: boolean
          description: Есть ли следующая страница
          example: true

    CartItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID записи в корзине
          example: 1
        productId:
          type: integer
          format: int64
          description: ID товара
          example: 1
        productTitle:
          type: string
          description: Название товара
          example: Мяч футбольный
        productImage:
          type: string
          description: URL изображения
          example: /images/football.jpg
        price:
          type: integer
          format: int64
          description: Цена
          example: 1000
        count:
          type: integer
          description: Количество
          example: 2
        sum:
          type: integer
          format: int64
          description: Сумма (price * count)
          example: 2000

    UpdateCartRequest:
      type: object
      required:
        - productId
        - action
      properties:
        productId:
          type: integer
          format: int64
          description: ID товара
          example: 1
        action:
          type: string
          enum: [PLUS, MINUS, DELETE]
          description: Действие (PLUS - добавить, MINUS - уменьшить, DELETE - удалить)
          example: PLUS

    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID заказа
          example: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalSum:
          type: integer
          format: int64
          description: Общая сумма заказа
          example: 231

    OrderItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID элемента заказа
          example: 1
        productTitle:
          type: string
          description: Название товара
          example: Мяч футбольный
        count:
          type: integer
          description: Количество
          example: 2
        price:
          type: integer
          format: int64
          description: Цена
          example: 7999900

    CreateOrderResponse:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
          description: ID созданного заказа
          example: 1

    CreateProductRequest:
      type: object
      required:
        - title
        - price
      properties:
        title:
          type: string
          description: Название товара
          minLength: 1
          example: Футбольный мяч
        description:
          type: string
          description: Описание товара
          example: Профессиональный мяч размер 5
        imgPath:
          type: string
          description: Путь к изображению
          example: /images/ball.jpg
        price:
          type: integer
          format: int64
          description: Цена товара
          minimum: 1
          example: 2500